requirement.txt
Django==1.11.6
django-classy-tags==0.8.0
django-cms==3.4.5
django-formtools==2.1
django-sekizai==0.10.0
django-treebeard==4.2.0
djangocms-admin-style==1.2.7
mysqlclient==1.3.12
pytz==2017.3
djangocms_text_ckeditor
aldryn_style
django-filer
djangocms_picture
-e git+https://github.com/aldryn/aldryn-newsblog.git#egg=aldryn-newsblog
djangorestframework
markdown
